SUBBOTNIK-4173 move library/testing/benchmark into library/cpp
ref:4a320e80735f5592483e5f14271226606920b8fc
