+++ catboost/libs/data/benchmarks/load_data_from_dsv_bench.cpp
@@ -4,7 +4,7 @@
 #include <catboost/libs/data/data_provider.h>
 #include <catboost/libs/data/objects_grouping.h>
 
-#include <library/testing/benchmark/bench.h>
+#include <library/cpp/testing/benchmark/bench.h>
 #include <library/cpp/unittest/tests_data.h>
 
 using namespace NCB;
