+++ library/cpp/digest/md5/bench/main.cpp
@@ -1,4 +1,4 @@
-#include <library/testing/benchmark/bench.h>
+#include <library/cpp/testing/benchmark/bench.h>
 #include <library/cpp/digest/md5/md5.h>
 
 #include <util/generic/xrange.h>
