+++ library/cpp/dot_product/bench/main.cpp
@@ -1,6 +1,6 @@
 #include <library/cpp/dot_product/dot_product.h>
 
-#include <library/testing/benchmark/bench.h>
+#include <library/cpp/testing/benchmark/bench.h>
 
 #include <util/generic/singleton.h>
 #include <util/generic/vector.h>
