+++ library/cpp/pop_count/benchmark/main.cpp
@@ -3,7 +3,7 @@
 #include <util/generic/type_name.h>
 
 #include <library/cpp/pop_count/popcount.h>
-#include <library/testing/benchmark/bench.h>
+#include <library/cpp/testing/benchmark/bench.h>
 
 template <class F, class I>
 inline void DoRun(F&& f, I&& i) {
