+++ library/cpp/threading/future/perf/main.cpp
@@ -1,4 +1,4 @@
-#include <library/testing/benchmark/bench.h>
+#include <library/cpp/testing/benchmark/bench.h>
 #include <library/cpp/threading/future/future.h>
 
 #include <util/generic/string.h>
