+++ library/cpp/threading/thread_local/benchmark/main.cpp
@@ -1,6 +1,6 @@
 #include <library/cpp/colorizer/colors.h>
 #include <library/cpp/statistics/statistics.h>
-#include <library/testing/benchmark/bench.h>
+#include <library/cpp/testing/benchmark/bench.h>
 #include <library/cpp/threading/thread_local/thread_local.h>
 
 #include <util/datetime/base.h>
