+++ library/cpp/threading/thread_local/benchmark/ya.make
@@ -9,7 +9,7 @@ PEERDIR(
     library/cpp/object_factory
     library/cpp/statistics
     library/cpp/threading/thread_local
-    library/testing/benchmark
+    library/cpp/testing/benchmark
 )
 
 END()
