--- library/testing/benchmark/bench.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <library/cpp/testing/benchmark/bench.h>
