--- library/testing/benchmark/main/ya.make
@@ -1,9 +1,0 @@
-LIBRARY()
-
-
-
-PEERDIR(
-    library/cpp/testing/benchmark/main
-)
-
-END()
