--- library/testing/benchmark/ya.make
@@ -1,13 +1,0 @@
-LIBRARY()
-
-
-
-SRCS(
-    bench.h
-)
-
-PEERDIR(
-    library/cpp/testing/benchmark
-)
-
-END()
