--- library/testing/ya.make
@@ -1,4 +1,0 @@
-RECURSE(
-    benchmark
-    benchmark/main
-)
