+++ util/charset/benchmark/to_lower/main.cpp
@@ -1,4 +1,4 @@
-#include <library/testing/benchmark/bench.h>
+#include <library/cpp/testing/benchmark/bench.h>
 
 #include <util/charset/wide.h>
 #include <util/generic/singleton.h>
