+++ util/generic/benchmark/sort/main.cpp
@@ -1,4 +1,4 @@
-#include <library/testing/benchmark/bench.h>
+#include <library/cpp/testing/benchmark/bench.h>
 
 #include <util/generic/algorithm.h>
 #include <util/generic/vector.h>
