+++ util/generic/benchmark/vector_count_ctor/f.cpp
@@ -1,6 +1,6 @@
 #include "f.h"
 
-#include <library/testing/benchmark/bench.h>
+#include <library/cpp/testing/benchmark/bench.h>
 
 #include <util/generic/vector.h>
 #include <util/generic/ptr.h>
