+++ util/random/benchmark/prng/main.cpp
@@ -1,4 +1,4 @@
-#include <library/testing/benchmark/bench.h>
+#include <library/cpp/testing/benchmark/bench.h>
 
 #include <util/random/entropy.h>
 #include <util/random/fast.h>
