+++ util/string/benchmark/join/main.cpp
@@ -1,4 +1,4 @@
-#include <library/testing/benchmark/bench.h>
+#include <library/cpp/testing/benchmark/bench.h>
 
 #include <util/generic/function.h>
 #include <util/generic/singleton.h>
