+++ util/string/benchmark/subst_global/main.cpp
@@ -1,4 +1,4 @@
-#include <library/testing/benchmark/bench.h>
+#include <library/cpp/testing/benchmark/bench.h>
 
 #include <util/generic/cast.h>
 #include <util/generic/singleton.h>
