+++ util/system/benchmark/cpu_id/main.cpp
@@ -1,4 +1,4 @@
-#include <library/testing/benchmark/bench.h>
+#include <library/cpp/testing/benchmark/bench.h>
 
 #include <util/system/cpu_id.h>
