+++ util/system/unaligned_mem_ut.cpp
@@ -1,6 +1,6 @@
 #include "unaligned_mem.h"
 
-#include <library/testing/benchmark/bench.h>
+#include <library/cpp/testing/benchmark/bench.h>
 #include <library/cpp/unittest/registar.h>
 
 #include <util/system/compiler.h>
